fixes search filter stack overflow

+++ src/components/FilterGroup/SearchFilter.tsx
@@ -1,7 +1,7 @@
 import { filterQueryAtom } from "@/atoms";
 import { useAtom } from "jotai";
 import { useEffect, useState } from "react";
-import { Stack, TextField } from "@mui/material";
+import { TextField } from "@mui/material";
 import { useDebounce } from "@uidotdev/usehooks";
 
 export default function SearchFilter({
@@ -10,10 +10,12 @@ export default function SearchFilter({
   title?: string;
 }) {
   const [filterQuery, setFilterQuery] = useAtom(filterQueryAtom);
-  const [searchTerm, setSearchTerm] = useState(filterQuery.get("search"));
+  const [searchTerm, setSearchTerm] = useState<string | null>(null);
   const debouncedTerm = useDebounce(searchTerm, 300);
 
   useEffect(() => {
+    if (debouncedTerm === null) return;
+
     const query = new URLSearchParams(filterQuery.toString());
     if (!debouncedTerm) {
       query.delete("search");
@@ -22,7 +24,7 @@ export default function SearchFilter({
     }
 
     setFilterQuery(query);
-  }, [debouncedTerm, setFilterQuery, filterQuery]);
+  }, [debouncedTerm]);
 
   return (
     <TextField
